Improved casting

+++ src/gol.cpp
@@ -5,9 +5,9 @@ void gol_step(Grid& grid) {
     for (uint32_t x = 0; x < grid_copy.get_width(); x++) {
         for (uint32_t y = 0; y < grid_copy.get_height(); y++) {
             uint8_t neighbors = 0;
-            for (uint32_t i = x - 1; i <= x + 1; i++) {
-                for (uint32_t j = y - 1; j <= y + 1; j++) {
-                    if (i >= 0 && i < grid_copy.get_width() && j >= 0 && j < grid_copy.get_height())
+            for (int32_t i = x - 1; i <= (int32_t)(x + 1); i++) {
+                for (int32_t j = y - 1; j <= (int32_t)(y + 1); j++) {
+                    if (i >= 0 && i < (int32_t)grid_copy.get_width() && j >= 0 && j < (int32_t)grid_copy.get_height())
                         if (grid_copy(i, j)) neighbors++;
                 }
             }
